# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nextenti Recruit CRM API",
    "version": "1.0.0",
    "description": "Backend API for the Nextenti Recruit CRM web app and integrations."
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nt_session"
      }
    },
...
  },
  "paths": {
...
    "/v1/queue": {
      "get": {
        "summary": "Outreach queue of Validated leads (own, or the whole team for the Team 1 leader)",
        "tags": [
          "queue"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "mine",
                "team"
              ]
            },
            "in": "query",
            "name": "scope",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            },
            "in": "query",
            "name": "overdue",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/missed-calls": {
      "get": {
        "summary": "Missed-call inbox (own, or the whole team for leaders) with this week's recall funnel",
        "tags": [
          "missed-calls"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "closed"
              ]
            },
            "in": "query",
            "name": "tab",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Log a missed incoming call (creates a recall task)",
        "tags": [
          "missed-calls"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mobile": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "receivedAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "mobile"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/availability": {
      "get": {
        "summary": "Qualified leads, due availability check-ins and cold → warm conversions",
        "tags": [
          "availability"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            },
            "in": "query",
            "name": "cold",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/scrutiny": {
      "get": {
        "summary": "Enrolled leads awaiting scrutiny, with SOP completeness",
        "tags": [
          "scrutiny"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "incomplete",
                "complete",
                "all"
              ]
            },
            "in": "query",
            "name": "tab",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/vacancies": {
      "get": {
        "summary": "Vacancies list with filters, sourcing stats and client-org options",
        "tags": [
          "vacancies"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 200
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 40
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 40
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 40
            },
            "in": "query",
            "name": "team",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "maxLength": 100
            },
            "in": "query",
            "name": "org",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false",
                "1",
                "0"
              ]
            },
            "in": "query",
            "name": "mine",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create a vacancy (routes to Team 3a/3b/3c by client org type)",
        "tags": [
          "vacancies"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clientOrgId": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "title": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "category": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "specialty": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "location": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "minExperienceYears": {
                    "type": "number"
                  },
                  "ctcMinLakhs": {
                    "type": "number"
                  },
                  "ctcMaxLakhs": {
                    "type": "number"
                  },
                  "maxNoticeDays": {
                    "type": "number"
                  },
                  "openings": {
                    "type": "number"
                  },
                  "postedAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/evaluations": {
      "get": {
        "summary": "Scorecards, newest first",
        "tags": [
          "evaluations"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": true,
              "minimum": 0
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "Create an evaluation comparing 1–3 candidates",
        "tags": [
          "evaluations"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "templateId": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "vacancyId": {
                    "type": "string",
                    "maxLength": 100,
                    "nullable": true
                  },
                  "candidateIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1
                    },
                    "maxItems": 50
                  }
                },
                "required": [
                  "candidateIds"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}